requirements.txt
aiohttp
asyncio
discord.py
plyer
pystray
pillow